+++ svalbard/xmlbeans/src/main/java/org/n52/svalbard/encode/AbstractTimeLocationValueTripleTypeEncoder.java
@@ -16,6 +16,7 @@
  */
 package org.n52.svalbard.encode;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 import org.apache.xmlbeans.XmlObject;
@@ -102,7 +103,7 @@ public abstract class AbstractTimeLocationValueTripleTypeEncoder<T>
         String value = null;
         if (timeLocationValueTriple.getValue() instanceof QuantityValue) {
             QuantityValue quantityValue = (QuantityValue) timeLocationValueTriple.getValue();
-            if (!quantityValue.getValue().equals(Double.NaN)) {
+            if (!quantityValue.getValue().equals(BigDecimal.valueOf(Double.NaN))) {
                 value = Double.toString(quantityValue.getValue().doubleValue());
             }
         } else if (timeLocationValueTriple.getValue() instanceof CountValue) {
